% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_helpers.R
\name{build_delete_query}
\alias{build_delete_query}
\title{Build DELETE query based on available parameters}
\usage{
build_delete_query(table_name, cohort_id, database_id, concept_id = NULL)
}
\description{
Build DELETE query based on available parameters
}
\keyword{internal}
