% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{correlation_analysis}
\alias{correlation_analysis}
\title{Correlation analysis for code recommendation when evaluating a phenotype}
\usage{
correlation_analysis(
  cohort_id,
  scratch,
  cdm_schema,
  concept_sets,
  conn,
  server,
  drivers_dir,
  max_cores = future::availableCores()\%/\%3L,
  filter_domains_to = c("condition_occurrence"),
  parallel_run = FALSE
)
}
\arguments{
\item{cohort_id}{atlas cohort id of the phenotype to be evaluated}

\item{scratch}{scratch space where the cohort is available}

\item{cdm_schema}{cdm schema name from the dataset}

\item{concept_sets}{concepts from the concept set with at least the following
columns (Id = concept_id, Name = concept_name, Domain = as retrieved from ATLAS)}
}
\description{
Create a list of tables with the result of the quantitative correlation
between the appearance of the codes from the concept set and the 30 most
frequent codes per domain between the 6 domains from the omop cdm (Condition,
Drug, Procedure, Device, Measurement, Observation).
}
\details{
\subsection{Required packages}{
\itemize{
\item data.table
}
}

\subsection{Required functions}{
\itemize{
\item table_freq_concepts
\item query_patientxconcept
}
}

\subsection{Other requirements}{
\itemize{
\item The connection to the dataset will have to be created before running the function.
}
}
}
\author{
Luisa Martínez (10SEP2024)
}
