% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{correlation_table_name}
\alias{correlation_table_name}
\title{correlation_table_name}
\usage{
correlation_table_name(table_name, coh_id, sep = "_")
}
\arguments{
\item{coh_id}{}
}
\description{
correlation_table_name
}
