% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concept_sets.R
\name{get_concept_sets}
\alias{get_concept_sets}
\title{get_concept_sets}
\usage{
get_concept_sets(
  included_source_concept.csv,
  concept.csv,
  threshold = .CONCEPTS_SETS_THRESHOLD,
  domain = "Condition",
  rename_cols = TRUE
)
}
\arguments{
\item{threshold}{confirmed with Martínez Sánchez, Luisa}
}
\value{
data.frame
}
\description{
get_concept_sets
}
