% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concept_sets.R
\name{get_concept_sets_from_db}
\alias{get_concept_sets_from_db}
\title{get_concept_sets_from_db}
\usage{
get_concept_sets_from_db(
  sqlite_path,
  threshold = .CONCEPTS_SETS_THRESHOLD,
  domain = "Condition",
  rename_cols = TRUE
)
}
\arguments{
\item{threshold}{}
}
\description{
get_concept_sets_from_db
}
