% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{get_correlation_analysis_output_table_names}
\alias{get_correlation_analysis_output_table_names}
\title{get_correlation_analysis_output_table_names}
\usage{
get_correlation_analysis_output_table_names(cohort_id = NA)
}
\description{
get_correlation_analysis_output_table_names
}
