% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{get_databricks_conn}
\alias{get_databricks_conn}
\title{get_databricks_conn}
\usage{
get_databricks_conn(
  drivers_dir = "drivers",
  user = "token",
  password = Sys.getenv("DATABRICKS_TOKEN"),
  databricks_host = Sys.getenv("DATABRICKS_HOST",
    "adb-7228388582267561.1.azuredatabricks.net"),
  databricks_http_path = Sys.getenv("DATABRICKS_HTTP_PATH",
    "/sql/1.0/warehouses/768ce28781fd35dc"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
get_databricks_conn
}
