% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_changes_analysis.R
\name{get_prevalence_changes_analysis_output_table_names}
\alias{get_prevalence_changes_analysis_output_table_names}
\title{Get prevalence changes analysis output table names}
\usage{
get_prevalence_changes_analysis_output_table_names()
}
\value{
character vector of table names
}
\description{
Returns the list of table names that are generated by the
prevalence changes analysis functions.
}
\seealso{
Other prevalence_changes_analysis: 
\code{\link{prevalence_changes_analysis}()},
\code{\link{read_prevalence_changes}()},
\code{\link{read_prevalence_changes_formatted}()}
}
\author{
Luisa Martínez (08APR2025)
}
\concept{prevalence_changes_analysis}
