% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{get_redshift_conn}
\alias{get_redshift_conn}
\title{get_redshift_conn}
\usage{
get_redshift_conn(
  server = Sys.getenv("redshift_server"),
  drivers_dir = "drivers",
  user = Sys.getenv("redshift_username"),
  password = Sys.getenv("redshift_password")
)
}
\arguments{
\item{password}{}
}
\description{
get_redshift_conn
}
