% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_dependent_analysis.R
\name{get_time_dependent_analysis_output_table_names}
\alias{get_time_dependent_analysis_output_table_names}
\title{Get time dependent analysis output table names}
\usage{
get_time_dependent_analysis_output_table_names()
}
\value{
character vector of table names
}
\description{
Returns the list of table names that are generated by the
time dependent analysis functions.
}
\seealso{
Other time_dependent_analysis: 
\code{\link{code_x_time}()},
\code{\link{patient_eofup}()},
\code{\link{patient_recruitment}()},
\code{\link{patient_x_time}()},
\code{\link{plot_counts}()},
\code{\link{plotly_counts}()},
\code{\link{read_all_time_dependent_tables}()},
\code{\link{read_time_dependent_table}()},
\code{\link{time_dependent_analysis}()}
}
\author{
Luisa Martínez (08APR2025)
}
\concept{time_dependent_analysis}
