% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{heatmap_plot}
\alias{heatmap_plot}
\title{Plot a heatmap from the tables retrieve using correlation_analysis}
\usage{
heatmap_plot(corr_information, table_domain)
}
\arguments{
\item{corr_information}{correlation table retrieve from the output of the function
correlation_analysis}

\item{table_domain}{name of the omop cdm domain represented in the table, this
name will appear in the title of the plot}
}
\description{
Plot a heatmap from the tables retrieve in a list from the
correlation_analysis function
}
\details{
\subsection{Required packages}{
\itemize{
\item ggplot2
\item plotly
\item stringr
}
}
}
