% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concept_sets.R
\name{merge_concept_sets}
\alias{merge_concept_sets}
\title{merge_concept_sets}
\usage{
merge_concept_sets(
  included_source_concept,
  concept,
  threshold = .CONCEPTS_SETS_THRESHOLD,
  domain = "Condition",
  rename_cols = TRUE
)
}
\arguments{
\item{domain}{}
}
\description{
merge_concept_sets
}
