% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_counts.R
\name{plotly_counts}
\alias{plotly_counts}
\title{plotly_counts}
\usage{
plotly_counts(
  table,
  time,
  counts,
  date_label = "Date",
  count_label = "Count",
  tit = "",
  title_font_size = 14
)
}
\description{
plotly_counts
}
\seealso{
Other time_dependent_analysis: 
\code{\link{code_x_time}()},
\code{\link{get_time_dependent_analysis_output_table_names}()},
\code{\link{patient_eofup}()},
\code{\link{patient_recruitment}()},
\code{\link{patient_x_time}()},
\code{\link{plot_counts}()},
\code{\link{read_all_time_dependent_tables}()},
\code{\link{read_time_dependent_table}()},
\code{\link{time_dependent_analysis}()}
}
\author{
Luisa Martínez (23MAY2025)
}
\concept{time_dependent_analysis}
