% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_visualization__query_patientxconcept.R
\name{query_all_patientxconcept}
\alias{query_all_patientxconcept}
\title{query_all_patientxconcept}
\usage{
query_all_patientxconcept(main_table, source_concept_id, concept_id)
}
\description{
Query the cdm to extract patients with a specific concept from any table
}
