% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_visualization__query_patientxconcept.R
\name{query_patientxconcept}
\alias{query_patientxconcept}
\title{Generate a query to extract the patients id with a concept from omop cdm}
\usage{
query_patientxconcept(
  domain_table,
  domain_concept_id,
  scratch,
  cdm_schema,
  concept_id,
  cohort_id
)
}
\arguments{
\item{domain_table}{name of the domain table from the omop cdm}

\item{domain_concept_id}{name of the field matching to concept_id from the omop cdm}

\item{scratch}{scratch space where the cohort is available}

\item{cdm_schema}{cdm schema name from the dataset}

\item{concept_id}{concept id from table concepts of omop cdm}

\item{cohort_id}{atlas cohort id of the phenotype to be evaluated}
}
\description{
Generate a query to extract the patients id with a specific concept
from a domain from the omop cdm
}
\author{
Luisa Martínez (10SEP2024)
}
