% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_freq_concepts.R
\name{table_freq_concepts}
\alias{table_freq_concepts}
\title{Generate a query to create a table with most frequent concepts per cohort/domain}
\usage{
table_freq_concepts(
  domain_table,
  domain_concept_id,
  scratch,
  cdm_schema,
  cohort_id,
  cut_off = 5
)
}
\arguments{
\item{domain_table}{name of the domain table from the omop cdm}

\item{domain_concept_id}{name of the field matching to concept_id from the omop cdm}

\item{scratch}{scratch space where the cohort is available}

\item{cdm_schema}{cdm schema name from the dataset}

\item{cohort_id}{atlas cohort id of the phenotype to be evaluated}

\item{cut_off}{minimum proportion of patients from the cohort having a code to
include the code into the output table}
}
\description{
Generate a query to create a table with the most frequent concepts
from a specific cohort in a specific domain from the scratch space .
}
\author{
Luisa Martínez (10SEP2024)
}
