% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_freq_spcf_concepts.R
\name{table_freq_spcf_concepts}
\alias{table_freq_spcf_concepts}
\title{Create a table with concept frequency in the general population}
\usage{
table_freq_spcf_concepts(
  domain_table,
  domain_concept_id,
  input_concepts,
  cdm_schema
)
}
\arguments{
\item{domain_table}{name of the domain table from the omop cdm}

\item{domain_concept_id}{name of the field matching to concept_id from the omop cdm}

\item{input_concepts}{concepts ids from table concepts of omop cdm to retrieved}

\item{cdm_schema}{cdm schema name from the dataset}
}
\description{
Create a table with concept_id and frequency of the general
population for specific given concepts.
}
\author{
Luisa Martínez (15SEP2024)
}
