% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_dependent_analysis.R
\name{time_dependent_analysis}
\alias{time_dependent_analysis}
\title{Run time dependent analysis for a cohort and save results to PostgreSQL}
\usage{
time_dependent_analysis(
  cohort_id,
  scratch.table,
  cdm_schema,
  concept_sets,
  conn,
  postgres_conn,
  database_id,
  server,
  drivers_dir,
  max_cores = 1,
  schema_name = "phenotype_library"
)
}
\arguments{
\item{cohort_id}{cohort identifier to analyze}

\item{scratch.table}{space and table where the cohort is available}

\item{cdm_schema}{general database cdm schema}

\item{concept_sets}{data frame containing concept sets information}

\item{conn}{database connection for querying data}

\item{postgres_conn}{PostgreSQL connection for saving results}

\item{database_id}{database identifier for the analysis}

\item{server}{server name for logging}

\item{drivers_dir}{directory containing database drivers}

\item{max_cores}{maximum number of cores to use for parallel processing}

\item{schema_name}{PostgreSQL schema name (default: "phenotype_library")}
}
\value{
list containing results from all four time dependent analysis functions
}
\description{
This function runs all four time dependent analysis functions
(patient_eofup, patient_recruitment, patient_x_time, code_x_time) for a given
cohort and saves the results to PostgreSQL database. This function is designed
to be called during cohort evaluation to populate the database with time
dependent analysis results.
}
\details{
\subsection{Required packages}{
\itemize{
\item data.table
\item DBI
\item PLNewFeatures
}
}

\subsection{Functions Called}{
\itemize{
\item patient_eofup: End of follow-up analysis
\item patient_recruitment: Patient recruitment analysis
\item patient_x_time: Patient count over time
\item code_x_time: Code frequency over time (for each concept in concept_sets)
}
}

\subsection{PostgreSQL Tables Created/Updated}{
\itemize{
\item patient_eofup: End of follow-up analysis results
\item patient_recruitment: Patient recruitment analysis results
\item patient_x_time: Patient count over time results
\item code_x_time: Code frequency over time results
}
}
}
\examples{
\dontrun{
# Run time dependent analysis for a cohort
results <- time_dependent_analysis(
  cohort_id = 123,
  scratch.table = "scratch.cohort_table",
  cdm_schema = "cdm_schema",
  concept_sets = concept_sets_df,
  conn = redshift_conn,
  postgres_conn = postgres_conn,
  database_id = "optum_ehr",
  server = "server_name",
  drivers_dir = "drivers",
  max_cores = 4
)
}

}
\seealso{
Other time_dependent_analysis: 
\code{\link{code_x_time}()},
\code{\link{get_time_dependent_analysis_output_table_names}()},
\code{\link{patient_eofup}()},
\code{\link{patient_recruitment}()},
\code{\link{patient_x_time}()},
\code{\link{plot_counts}()},
\code{\link{plotly_counts}()},
\code{\link{read_all_time_dependent_tables}()},
\code{\link{read_time_dependent_table}()}
}
\author{
Luisa Martínez (08APR2025)
}
\concept{time_dependent_analysis}
