% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{validate_db_table_perf}
\alias{validate_db_table_perf}
\title{validate_db_table_perf}
\usage{
validate_db_table_perf(string)
}
\arguments{
\item{string}{}
}
\value{
logical
}
\description{
validate_db_table_perf
}
